improv: make timeline page protected behind login
